prevent programmatic custom block deletion from scenes with a puzzle solution

+++ src/threads.js
@@ -65,7 +65,7 @@ StagePickerMorph, CustomBlockDefinition, CommentMorph*/
 
 /*jshint esversion: 11, bitwise: false, evil: true*/
 
-modules.threads = '2023-July-12';
+modules.threads = '2023-July-14';
 
 var ThreadManager;
 var Process;
@@ -8115,6 +8115,11 @@ Process.prototype.doDeleteBlock = function (context) {
     expr = context.expression;
     if (!expr.isCustomBlock) {
         throw new Error('expecting a custom block\nbut getting a primitive');
+    }
+    if (ide.sprites.asArray().some(any => any.solution)) {
+        throw new Error(
+            'cannot delete custom blocks\nfrom a scene with a puzzle solution'
+        );
     }
     def = expr.isGlobal ? expr.definition : rcvr.getMethod(expr.semanticSpec);
     rcvr.deleteAllBlockInstances(def);
